fix: move UiShell, renderReactInWebComponent into separate export for React 17 compatibility
