+++ packages/odyssey-react-mui/package.json
@@ -33,6 +33,10 @@
       "types": "./dist/src/labs/index.d.ts",
       "default": "./dist/labs/index.js"
     },
+    "./ui-shell": {
+      "types": "./dist/src/ui-shell/index.d.ts",
+      "default": "./dist/ui-shell/index.js"
+    },
     "./test-selectors": {
       "types": "./dist/src/test-selectors/index.d.ts",
       "default": "./dist/test-selectors/index.js"
