+++ packages/odyssey-react-mui/src/labs/index.ts
@@ -44,4 +44,3 @@ export * from "./AppSwitcher";
 export * from "./SideNav/NavAccordion";
 export * from "./SideNav";
 export * from "./TopNav";
-export * from "./UiShell";
