+++ packages/odyssey-react-mui/src/ui-shell/index.ts
@@ -1,5 +1,5 @@
 /*!
- * Copyright (c) 2024-present, Okta, Inc. and/or its affiliates. All rights reserved.
+ * Copyright (c) 2022-present, Okta, Inc. and/or its affiliates. All rights reserved.
  * The Okta software accompanied by this notice is provided pursuant to the Apache License, Version 2.0 (the "License.")
  *
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
@@ -10,16 +10,5 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import { useEffect, useState } from "react";
-
-import { uiShellDataAttribute } from "./renderUiShell";
-
-export const useHasUiShell = () => {
-  const [hasUiShell, setHasUiShell] = useState(false);
-
-  useEffect(() => {
-    setHasUiShell(Boolean(document.querySelector(`[${uiShellDataAttribute}]`)));
-  }, []);
-
-  return hasUiShell;
-};
+export * from "./UiShell";
+export * from "./renderReactInWebComponent";
